allow sql inputs to be either strings of sql or filenames to read
If the string does not contain a semicolon or a space, it is not valid sql
and is instead treated as a file name and read with mu_read_small_file.

To prevent a lot of functional code duplication in mu_query3() and mu_query()
I decided to move to a two-part query where queries must first be created
with mu_create_query(), which will return a pointer to a new struct mu_QUERY,
and then run with mu_run_query3() or mu_run_query() which replace the old
functions mu_query3 and mu_query() respectively.